Implement CRUD form with four product filters,separated views for admin and guests

+++ backend/RS1_2024_25.API/Helper/Api/ProductController.cs
@@ -37,18 +37,22 @@ namespace RS1_2024_25.API.Helper.Api
 
         [HttpPost]
        [Authorize(Roles = "Admin")]
-        public async Task<IActionResult> CreateProduct([FromBody] ProductUpdateDTO dto)
+        public async Task<IActionResult> CreateProduct([FromBody] ProductCreateDTO dto)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
             var product = new Product
             {
-                Name=dto.Name,
-                Price = dto.Price ?? 0,
+                Name = dto.Name,
+                SKU = dto.SKU,
+                Price = dto.Price,
                 StockQuantity = dto.StockQuantity,
                 Description = dto.Description,
                 ImageUrl = dto.ImageUrl,
+                CategoryId = dto.CategoryId,
+                Brend = dto.Brend,
+                Active = dto.Active
 
             };
 
@@ -66,14 +70,14 @@ namespace RS1_2024_25.API.Helper.Api
                 return NotFound();
 
           if(dto.Name!=null)product.Name = dto.Name;
-            if (dto.Price.HasValue) product.Price = dto.Price.Value;
+           if(dto.Price.HasValue) product.Price = dto.Price.Value;
             if (dto.StockQuantity.HasValue) product.StockQuantity = dto.StockQuantity.Value;
             if (dto.Description != null) product.Description = dto.Description;
             if (dto.ImageUrl != null) product.ImageUrl = dto.ImageUrl;
             if(dto.Active!=null)product.Active = dto.Active.Value;
             if (dto.SKU != null) product.SKU = dto.SKU;
             if (dto.Brend != null) product.Brend = dto.Brend;
-            if (dto.CategoryId != null) product.CategoryId = dto.CategoryId;
+            if(dto.CategoryId.HasValue) product.CategoryId = dto.CategoryId.Value;
             if (dto.AdditionalImagesUrl != null) product.AdditionalImagesUrl = dto.AdditionalImagesUrl;
 
 
@@ -89,10 +93,31 @@ namespace RS1_2024_25.API.Helper.Api
             if (product == null)
                 return NotFound();
 
+
+            if (_context.OrderItems.Any(o => o.ProductId == id))
+                return BadRequest("Ne možete obrisati proizvod koji je dio neke narudžbe.");
+
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
             return Ok(new { Message = "Product deleted successfully." });
         }
+
+        [HttpGet("images")]
+        public IActionResult GetImages()
+        {
+            string imageFolderPath = Path.Combine(Directory.GetCurrentDirectory(),"wwwroot", "images","products");
+
+            if (!Directory.Exists(imageFolderPath))
+                return NotFound("Folder Images doesnt exist.");
+
+            var files = Directory.GetFiles(imageFolderPath)
+                .Select(Path.GetFileName)
+               
+                .ToList();
+
+            return Ok(files);
+        }
+
     }
 
 }
